Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
authlib==1.4.0
itsdangerous==2.2.0
python-dotenv==1.0.1
httpx>=0.24.0,<0.28.0
psycopg2-binary==2.9.11
nude>=0.1.0
supabase==2.0.0
